feat(e2e): replace provider redirection test with DEX native swap flow
- Replace `runSwapCheckProvider` with `runSwapDexNativeFlow` that tests
  the full DEX native flow (select provider → execute → approve → sign)
- Extract shared setup helpers into `swap.other.setup.ts`
- Export `getAmountFromUSD` from swap utils and add optional `provider`
  param to `getMinimumSwapAmount` for provider-scoped queries
- Remove unused `verifyLiveAppTitle` and `liveAppTitle` from swapLiveApp
- Update OneInch and Velora specs to use the new flow
